Added test command to test the server connection
The test command will start by connecting to the provided server and
authenticate the provided user via a user token. As this server doesn't
verify the token yet it will allow any user to connect.

Capcoms server on the other hand expects the steamid64 in hex as a
user name/id and an encrypted session ticket for DDDA. These can
currently only be aquired by running the Dune.Emulator .NET/Windows
client, which will aquire both it via Steam API. This of course requires
the user to be logged into the Steam client and own DDDA.

- Also improved error handing when receiving unexpected packets.
- Also fixed serializing and deserializing DataChunkPackets as they contain chunk lengths twice for some reason.